requirements.txt
Flask==2.3.0
Werkzeug<3
python-dotenv==1.0.0
pytest==7.4.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors>=4.0
authlib>=1.3
requests>=2.31
anthropic>=0.25
stripe>=7.0
resend>=2.0
gunicorn>=21.0
Flask-Limiter>=3.5
psycopg2-binary>=2.9
itsdangerous>=2.1
